requirements.txt
flask 
python-telegram-bot 
python-dotenv 
aiohttp
